State prompter
Added prompt that asks for state/territory and returns alert of state, capital, and abbreviation. However, runs infinitely. I'd like to find a way to incorporate a button that begins the prompt and have the alert end the function, so that you would have to click the button again to call the function askAboutState(). Left numberCheck() in so that I could check that I had the right number of states, abbreviations, and capitals so the array indexes would work when askAboutState() is called.