Removed redundant guesses
Previously one could sink a ship by hitting the same location on a ship 3 times, as all that was required for a sink was 3 hits, not a hit in each location. Added an if statement in the fire function that will return a message if attempting to fire on an already hit location of a ship.