Breakout
An incomplete breakout game created following a tutorial by Mt. Ford Studios at https://youtu.be/7UQYG5x9rZo. Currently, the ball will bounce off of the walls or paddle by simply reversing the sign of the relevant velocity. If it hits a horizontal wall, ball.xv = -ball.xv, and so on. This does not allow much control to the player, and Mt. Ford Studios' solution is to add an element of randomness to the ball bounce when it hits the paddle. However, I think that it would be better to apply "spin" to the ball based on its relative position on the paddle when they collide, so that a collision with the far left of the paddle bounces the ball left at a low angle, a collision dead center bounces the ball directly vertical, etc. This commit is prior to the application of either technique.