reactive bounce
ball bounce from lines 212-223 not working as intended, angle seems random and ball sometimes seems to pass through paddle. Intention was to have ball bounce relative to where it hits the paddle by applying a multiplier based on distance from center of paddle to the velocity, but this is not functioning properly.