working on powerups
made ball round for visual preference, working on sticky paddle powerup. Currently, ball does not release vertically and moves horizontally along paddle slightly. This may be because the framerate does not allow the ball to get far enough away before the sticky paddle recognizes that ball.y is touching the paddle, so it sticks again. Working on this.