Fixed sticky paddle powerup
Sticky paddle powerup was not not releasing ball because the function used to bounce the ball relative to where it connects with the paddle did not allow for 0 as a y velocity, since it was not intended to be used for serves. Bounce function worked based on 
(ball.yv > 0 ? -1 : 1)
in order to reverse the sign when the ball connected with paddle and send it back up. However, 0 is not > 0, so it did not reverse the sign and the ball wanted to continue moving down through the paddle but couldn't on a serve. Fixed by changing > to >=, which now appears to function as intended.