Terrain generation w/ noise
Simulates flying over randomly generated terrain, with extremeness of terrain adjustable by adjusting noise (xoff and yoff). Converted to JS with p5.js from The Coding Train's video on this in Processing (https://youtu.be/IKB1hWWedMk). This version looks smooth while the processing version shows all vertices, which would be preferred here. However, this is smooth because noFill() and stroke() functions don't appear to work the way one would expect, and the vertices don't display, so a solid colored fill was preferable to the solid white that was displaying during attempts to show vertices.