Updates
Changed color scheme for personal preference, made Sound function for audio playback so volume could be controlled, as it was very loud otherwise. Added random music that plays upon ball serve. Choices are stored in an array, and a random index is selected on a serve--however only one song is selected at a time due to the method of selection. It will remain that song until the window is refreshed--planning to work on this later so the song would change each level/game over