Audio Refactor, powerup work
Refactored audio to be handled through Howler JS library. Previous vanilla JS handling of audio is still in code, but commented out. Added new BigBall powerup that makes ball larger. Began working on adding a Floor powerup, which would effectively add one large brick under the paddle as a "safety net" to bounce the ball back one time if the paddle ever misses the ball. Implementation was not working as intended, but progress is in code in the form of comments.